% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imcid.fun.R
\name{g.ifun}
\alias{g.ifun}
\title{The square of the score function for determining the individualized MCID}
\usage{
g.ifun(z_tilde, val, ind1, ind2, w)
}
\description{
The square of the score function for determining the individualized MCID
}
\keyword{internal}
