% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcid.fun.R
\name{g.pfun}
\alias{g.pfun}
\title{The square of the score function for determining the population MCID}
\usage{
g.pfun(val, ind1, ind2, w)
}
\description{
The square of the score function for determining the population MCID
}
\keyword{internal}
