% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imcid.fun.R
\name{h.ifun}
\alias{h.ifun}
\title{The hassen matrix function for determining the individualized MCID}
\usage{
h.ifun(z_tilde, ind1, ind2, w)
}
\description{
The hassen matrix function for determining the individualized MCID
}
\keyword{internal}
