% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcid.fun.R
\name{h.pfun}
\alias{h.pfun}
\title{The hassen matrix function for determining the population MCID}
\usage{
h.pfun(ind1, ind2, w)
}
\description{
The hassen matrix function for determining the population MCID
}
\keyword{internal}
