% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imcid.fun.R
\name{imcid.hinge.smooth}
\alias{imcid.hinge.smooth}
\title{The value function used to search the initial value for determining the individualized MCID}
\usage{
imcid.hinge.smooth(x, y, z, w, n, parm, lambda, delta)
}
\description{
The value function used to search the initial value for determining the individualized MCID
}
\keyword{internal}
