% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imcid.fun.R
\name{imcid.ramp.smooth}
\alias{imcid.ramp.smooth}
\title{The value function needed to be optimized for determining the individualized MCID}
\usage{
imcid.ramp.smooth(x, y, z, w, n, parm, lambda, delta, t)
}
\description{
The value function needed to be optimized for determining the individualized MCID
}
\keyword{internal}
