% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imcidcv.smooth.R
\name{imcidcv.smooth}
\alias{imcidcv.smooth}
\title{The function used to determine the value of target function for different value of delta in the individualized MCID setting}
\usage{
imcidcv.smooth(x, y, z, lambda, delta, fold, maxit, tol)
}
\description{
The function used to determine the value of target function for different value of delta in the individualized MCID setting
}
\keyword{internal}
