% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcid.fun.R
\name{pmcid.ramp.smooth}
\alias{pmcid.ramp.smooth}
\title{The value function needed to be optimized for determining the population MCID}
\usage{
pmcid.ramp.smooth(x, y, w, n, tau, t, delta)
}
\description{
The value function needed to be optimized for determining the population MCID
}
\keyword{internal}
