% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcidcv.smooth.R
\name{pmcidcv.smooth}
\alias{pmcidcv.smooth}
\title{The function used to determine the value of target function for different value of delta in the population MCID setting}
\usage{
pmcidcv.smooth(x, y, delta, fold, maxit, tol)
}
\description{
The function used to determine the value of target function for different value of delta in the population MCID setting
}
\keyword{internal}
